% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{bounds_fit}
\alias{bounds_fit}
\title{Causal inference with multiple versions of control}
\usage{
bounds_fit(Y0, Y1, D0, D1, alpha_level)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{D1}{observed treatment types in the Z = 1 arm}

\item{alpha_level}{the significance level at which confidence
intervals should be constructed}
}
\value{
a 4 x 4 matrix containing the bounds estimates, standard errors
and upper and lower confidence bounds
}
\description{
Finds upper and lower bounds for the unidentified parameters
along with confidence intervals for the identified region
}

