% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{bounds_grad}
\alias{bounds_grad}
\title{Gradient of bounds function}
\usage{
bounds_grad(Y0, D0, Y1, D1)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D1}{observed treatment types in the Z = 1 arm}
}
\value{
a 4 x 7 matrix of partial derivatives of the upper and lower bounds
with respect to the observed probabilites q_dyz
}
\description{
Gradient of bounds function
}

