% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{bounds_vcov}
\alias{bounds_vcov}
\title{Asymptotic covariance matrix of the upper and lower bounds estimators}
\usage{
bounds_vcov(Y0, D0, Y1, D1)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D1}{observed treatment types in the Z = 1 arm}
}
\value{
a 4 x 4 estimated variance-covariance matrix
}
\description{
Asymptotic covariance matrix of the upper and lower bounds estimators
}

