% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{cace_bounds_grad}
\alias{cace_bounds_grad}
\title{Gradient of upper and lower bounds on CACE}
\usage{
cace_bounds_grad(Y0, D0, Y1, D1)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D1}{observed treatment types in the Z = 1 arm}
}
\value{
a 2 x 8 matrix of partial derivatives of the upper and lower bounds
with respect to the observed probabilites q_dyz
}
\description{
Gradient of upper and lower bounds on CACE
}

