% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{cace_vcov}
\alias{cace_vcov}
\title{Asymptotic covariance matrix for upper and lower bounds on CACE}
\usage{
cace_vcov(Y0, Y1, D0, D1)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{D1}{observed treatment types in the Z = 1 arm}
}
\value{
a 2 x 2 covariance matrix for the upper and lower bounds on the CACE
}
\description{
Asymptotic covariance matrix for upper and lower bounds on CACE
}

