% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{get_bounds}
\alias{get_bounds}
\title{Find upper and lower bounds for P(Y(1) | D(0) = 0, D(1) = 1) and
P(Y(1) | D(0) = 2, D(1) = 1)}
\usage{
get_bounds(Y0, Y1, D0, D1)
}
\arguments{
\item{Y0}{observed outcomes in Z = 0 arm}

\item{Y1}{observed outcomes in Z = 1 arm}

\item{D0}{observed treatment types in D = 0 arm}

\item{D1}{observed treatment types in D = 1 arm}
}
\value{
a 4-element vector of upper and lower bounds estimates
}
\description{
Find upper and lower bounds for P(Y(1) | D(0) = 0, D(1) = 1) and
P(Y(1) | D(0) = 2, D(1) = 1)
}

