% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{get_cace}
\alias{get_cace}
\title{Find upper and lower bounds for the CACE}
\usage{
get_cace(Y0, Y1, D0, D1)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{D1}{observed treatment types in the Z = 1 arm}
}
\value{
a 2-element vector containing the upper and lower bounds for the CACE
}
\description{
Find upper and lower bounds for the CACE
}

