% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{get_observedprob}
\alias{get_observedprob}
\title{Observed data quantities}
\usage{
get_observedprob(Y0, D0, Y1, D1)
}
\arguments{
\item{Y0}{observed outcomes in Z = 0 arm}

\item{D0}{observed treatment types in Z = 0 arm}

\item{Y1}{observed outcomes in Z = 1 arm}

\item{D1}{observed treatment types in Z = 1 arm}
}
\description{
Calculates q_dyz = P(Y = y, D = d | Z = z)
}

