% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{model_test}
\alias{model_test}
\title{Testing model assumptions}
\usage{
model_test(Y0, Y1, D0, D1)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{D1}{observed treatment types in the Z = 1 arm}
}
\value{
a logical, TRUE indicates that the data are consistent with the assumptions,
FALSE indicates that the data violate at least one of the assumptions.
}
\description{
\code{model_test} checks whether the data violate the assumptions underlying the model,
similar to the assumptions tests for the instrumental variable model.
}

