% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{observedprob_vcov}
\alias{observedprob_vcov}
\title{Asymptotic covariance matrix for estimatates of P(Y = y, D = d)}
\usage{
observedprob_vcov(Y, D, Z)
}
\arguments{
\item{Y}{vector of observed outcomes}

\item{D}{vector of observed treatment types}

\item{Z}{indicator for which treatment arm the data come from}
}
\value{
a matrix of variance-covariance estimates
}
\description{
Asymptotic covariance matrix for estimatates of P(Y = y, D = d)
}

