% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleversions_code_110316.R
\name{pointident_fit}
\alias{pointident_fit}
\title{Inference for Point Identified Quantities}
\usage{
pointident_fit(Y0, D0, Y1, D1, alpha_level)
}
\arguments{
\item{Y0}{observed outcomes in the Z = 0 arm}

\item{D0}{observed treatment types in the Z = 0 arm}

\item{Y1}{observed outcomes in the Z = 1 arm}

\item{D1}{observed outcomes in the Z = 1 arm}
}
\value{
a matrix containing point estimates, standard errors
and confidence intervals
}
\description{
Finds point estimates, standard errors and confidence intervals
for point identified quantites in the multiple versions framework
}

